refactor(structure): reorganize components into shared and pages directories
