+++ src/app/app.component.scss
@@ -1,4 +1,4 @@
-@import './styles/variables';
+@use './styles/variables' as vars;
 
 :host {
   display: flex;
@@ -11,5 +11,5 @@
   content: '';
   display: block;
   height: 100px; // Approximate height of the footer
-  margin-bottom: $spacing-md;
+  margin-bottom: vars.$spacing-md;
 }
