+++ src/app/pages/groupes/groupes.component.html
@@ -1,6 +1,6 @@
 <div class="user-dashboard">
- 
-  <h1 class="dashboard-title">User Dashboard</h1>
+
+  <h1 class="dashboard-title">Groupes</h1>
 
   <!-- Affichage de l'alerte en cas d'erreur -->
   <app-alert
@@ -23,7 +23,7 @@
     <!-- Carte "My Groups" -->
     <div class="my-groups-card">
       <div class="card-header">My Groups</div>
-      
+
       <div class="card-body">
         <div class="groups-count">{{ totalGroups }}</div>
         <div class="groups-label">groupes</div>
