+++ src/app/pages/listDetails/listDetails.html
