+++ src/app/pages/listDetails/listDetails.scss
