+++ src/app/pages/listDetails/listDetails.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterLink } from '@angular/router';
-import { ButtonComponent }           from '../../../shared/components/button/button.component';
+import { ButtonComponent }           from '../../shared/components/button/button.component';
 
 
 @Component({
