+++ src/app/shared/components/header/header.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { Router, ActivatedRoute, NavigationEnd } from '@angular/router';
 import { filter, map } from 'rxjs/operators';
-import { BackButtonComponent } from '../components/back.button/back.button.component';
+import { BackButtonComponent } from '../back.button/back.button.component';
 
 @Component({
   selector: 'app-header',
